test(utils): moved tests in spec folder
