+++ src/utils/spec/command.spec.ts
@@ -1,5 +1,5 @@
 import { Logger, LogLevel } from '@sasjs/utils/logger'
-import { Command } from './command'
+import { Command } from '../command'
 
 describe('parseCommandLine', () => {
   const defaultFlagNames = ['target']
